Login Form Creation
Created the login page...

No way to test.
:(

But created a auth page to check and see if the users is an actual user and has an account.

Also added  the Login form and used axios to connect the backend. Need the backend information to actually connect.

I also create a button to create an account on the form and it WILL send to the register from once i know the correct component to send it to.

Set up some error message as well based on login issues.
Server issue, login failed, missing username and password, and Unauthorized if we end up doing an admin